refactor: CardAcceptanceTest 및 DomainAcceptanceTest 리팩토링
