+++ backend/src/test/java/botobo/core/documentation/CardDocumentationTest.java
@@ -15,7 +15,6 @@ import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
-import java.util.Arrays;
 import java.util.List;
 
 import static botobo.core.documentation.DocumentationUtils.getDocumentRequest;
