+++ backend/src/test/java/botobo/core/quiz/CardAcceptanceTest.java
@@ -1,24 +1,20 @@
 package botobo.core.quiz;
 
 import botobo.core.DomainAcceptanceTest;
-import botobo.core.auth.AuthAcceptanceTest;
 import botobo.core.quiz.dto.NextQuizCardsRequest;
-import io.restassured.RestAssured;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
+import botobo.core.utils.RequestBuilder.HttpResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 import java.util.Arrays;
 import java.util.List;
 
-import static botobo.core.Fixture.CARD_REQUEST_1;
-import static botobo.core.Fixture.CARD_REQUEST_2;
-import static botobo.core.Fixture.CARD_REQUEST_3;
-import static botobo.core.Fixture.WORKBOOK_REQUEST_1;
+import static botobo.core.utils.Fixture.CARD_REQUEST_1;
+import static botobo.core.utils.Fixture.CARD_REQUEST_2;
+import static botobo.core.utils.Fixture.CARD_REQUEST_3;
+import static botobo.core.utils.Fixture.WORKBOOK_REQUEST_1;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("카드 인수 테스트")
@@ -39,16 +35,12 @@ public class CardAcceptanceTest extends DomainAcceptanceTest {
                 .build();
 
         // when
-        final ExtractableResponse<Response> response =
-                RestAssured.given().log().all()
-                        .contentType(MediaType.APPLICATION_JSON_VALUE)
-                        .auth().oauth2(로그인되어_있음().getAccessToken())
-                        .body(request)
-                        .when().put("/api/cards/next-quiz")
-                        .then().log().all()
-                        .extract();
+        final HttpResponse response = request()
+                .put("/api/cards/next-quiz", request)
+                .auth()
+                .build();
 
         // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT);
     }
 }
